Move configuration.py to _internal
