+++ certbot-compatibility-test/certbot_compatibility_test/configurators/nginx/common.py
@@ -7,7 +7,7 @@ import zope.interface
 
 from acme.magic_typing import Set  # pylint: disable=unused-import, no-name-in-module
 
-from certbot import configuration
+from certbot._internal import configuration
 from certbot_nginx import configurator
 from certbot_nginx import constants
 from certbot_compatibility_test import errors
