+++ certbot/_internal/main.py
@@ -18,7 +18,7 @@ from certbot._internal import account
 from certbot._internal import cert_manager
 from certbot import cli
 from certbot._internal import client
-from certbot import configuration
+from certbot._internal import configuration
 from certbot import constants
 from certbot import crypto_util
 from certbot._internal import eff
