+++ certbot/tests/main_test.py
@@ -23,7 +23,7 @@ from acme.magic_typing import List  # pylint: disable=unused-import, no-name-in-
 import certbot.tests.util as test_util
 from certbot._internal import account
 from certbot import cli
-from certbot import configuration
+from certbot._internal import configuration
 from certbot import constants
 from certbot import crypto_util
 from certbot import errors
