+++ certbot/tests/util.py
@@ -19,7 +19,7 @@ from six.moves import reload_module  # pylint: disable=import-error
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import serialization
 
-from certbot import configuration
+from certbot._internal import configuration
 from certbot import constants
 from certbot import interfaces
 from certbot._internal import lock
