--- docs/api/configuration.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.configuration`
---------------------------------
-
-.. automodule:: certbot.configuration
-   :members:
